only disable the backup mechanism on darwin
