+++ cmd/state-installer/internal/installer/installation.go
@@ -85,29 +85,14 @@ func (i *Installation) RemoveBackupFiles() error {
 	return nil
 }
 
-func (i *Installation) BackupFiles() error {
-	// Todo: https://www.pivotaltracker.com/story/show/177600107
-	// Get target file paths.
-	var targetFiles []string
-	for _, file := range fileutils.ListDir(i.fromDir, false) {
-		targetFile := filepath.Join(i.binaryDir, filepath.Base(file))
-		targetFiles = append(targetFiles, targetFile)
-	}
-	logging.Debug("Target files=%s", strings.Join(targetFiles, ","))
-
-	backups, err := backupFiles(targetFiles)
-	if err != nil {
-		return errs.Wrap(err, "Backup of existing files failed.")
-	}
-	i.backups = append(i.backups, backups...)
-	return nil
-}
-
 func (i *Installation) Rollback() error {
 	return restoreFiles(i.backups)
 }
 
 func (i *Installation) Install() error {
+	if err := i.BackupFiles(); err != nil {
+		return errs.Wrap(err, "Could not create backup files")
+	}
 	if err := fileutils.MkdirUnlessExists(i.binaryDir); err != nil {
 		return errs.Wrap(err, "Could not create target directory: %s", i.binaryDir)
 	}
